chatGroupMessage

+++ app/Http/Controllers/User/UserController.php
@@ -5,8 +5,10 @@ namespace App\Http\Controllers\User;
 use Illuminate\Http\Request;
 use App\Http\Controllers\Controller;
 use App\Http\Requests\LoginRequest;
+use App\Http\Requests\MessageRequest;
 use App\Services\UserServices;
 use App\Repositories\Contracts\UserRepository;
+use App\Model\User;
 
 class UserController extends Controller
 {
@@ -21,36 +23,58 @@ class UserController extends Controller
 
     public function postEditUser(LoginRequest $request, $id)
     {
-        $this->userServices->updateUser($request, $id);
-        $user = $this->userRepository->find($request->id);
-        session()->flash('success', 'admin.success');
+        $user = auth()->user();
+        $check = $this->userRepository->find($id);
+        if ($user->can('update', $check)) {
+            $this->userServices->updateUser($request, $id);
+            $user = $this->userRepository->find($request->id);
+            session()->flash('success', 'admin.success');
 
-        return response()->json($user);
+            return response()->json($user);
+        } else {
+            return 'ok';
+        }
     }
 
     public function postDeleteUser($id)
     {
-        $this->userServices->deleteUser($id);
-        session()->flash('success', 'success');
+        $user = auth()->user();
+        $check = $this->userRepository->find($id);
+        if ($user->can('delete', $check)) {
+            $this->userServices->deleteUser($id);
+            session()->flash('success', 'success');
+        } else {
+            return 'ok';
+        }
     }
 
     public function postAddUser(LoginRequest $request)
     {
-        $dataUser = [
-            'name' => $request->name,
-            'email' => $request->email,
-            'password' => $request->password,
-        ];
-        $this->userRepository->create($dataUser);
-        $maxUser = $this->userRepository->maxUser();
-        session()->flash('success', 'success');
-
-        return response()->json($maxUser);
+        $user = auth()->user();
+        if ($user->can('create', User::class)) {
+            $dataUser = [
+                'name' => $request->name,
+                'email' => $request->email,
+                'password' => $request->password,
+            ];
+            $this->userRepository->create($dataUser);
+            $maxUser = $this->userRepository->maxUser();
+            session()->flash('success', 'success');
+            return response()->json($maxUser);
+        } else {
+            return 'ok';
+        }
     }
 
     public function getAvatar($id)
     {
         $avatar = $this->userRepository->getAvatar($id);
         return $avatar;
+    }
+
+    public function searchUser(Request $request)
+    {
+        $user = $this->userRepository->searchUser($request->text);
+        return response()->json($user);
     }
 }
